Rename files and rename some functions and variables
